Updated to Gradle
Better spaces, removed unused events, changed to gradle, new plugin.yml, command if improved, gitignore
