--- .idea/artifacts/PteroBungee.xml
@@ -1,8 +1,0 @@
-<component name="ArtifactManager">
-  <artifact type="jar" name="PteroBungee">
-    <output-path>$PROJECT_DIR$/out/artifacts/PteroBungee</output-path>
-    <root id="archive" name="PteroBungee.jar">
-      <element id="module-output" name="BungeeStuff" />
-    </root>
-  </artifact>
-</component>
