--- .idea/libraries/BungeeCord.xml
@@ -1,9 +1,0 @@
-<component name="libraryTable">
-  <library name="BungeeCord">
-    <CLASSES>
-      <root url="jar://$PROJECT_DIR$/../../../../Descargas/BungeeCord.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES />
-  </library>
-</component>
