+++ build.gradle
@@ -1,0 +1,19 @@
+plugins {
+    id 'java'
+}
+
+group = 'us.sparkedhost'
+version = '1.2.0'
+
+repositories {
+    mavenCentral()
+    maven { url "https://oss.sonatype.org/content/repositories/snapshots" }
+}
+
+dependencies {
+    testImplementation 'org.junit.jupiter:junit-jupiter-api:5.6.0'
+    compileOnly 'net.md-5:bungeecord-api:1.16-R0.4'
+}
+
+sourceCompatibility = '1.8'
+targetCompatibility = '1.8'
