--- src/plugin.yml
@@ -1,5 +1,0 @@
-name: PteroBungee
-main: me.bettd.pterobungee.Main
-description: Plugin aims to improve broken compatibility of Pterodactyl panel with BungeeCord
-version: 1.1.0
-author: BetTD
